标尺刻度绘制

+++ app/src/main/java/com/john/ankotlin/DisplayUtils.java
@@ -37,7 +37,7 @@ public class DisplayUtils {
      * @param dpValue
      * @return
      */
-    public static int dp2px(Context context, float dpValue) {
+    public static float dp2px(Context context, float dpValue) {
         final float scale = context.getResources().getDisplayMetrics().density;
         return (int) (dpValue * scale + 0.5f);
     }
